+++ vscode/src/custom-prompts/CommandsController.ts
@@ -9,7 +9,7 @@ import {
 import { VsCodeCommandsController } from '@sourcegraph/cody-shared/src/editor'
 import { TelemetryService } from '@sourcegraph/cody-shared/src/telemetry'
 
-import { debug } from '../log'
+import { logDebug, logError } from '../log'
 import { LocalStorage } from '../services/LocalStorageProvider'
 
 import { CustomPromptsStore } from './CustomPromptsStore'
@@ -113,7 +113,7 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
     public find(id: string, isSlash = false): string {
         const myPrompt = this.default.get(id, isSlash)
 
-        debug('CommandsController:command:finding', id, { verbose: myPrompt })
+        logDebug('CommandsController:command:finding', id, { verbose: myPrompt })
 
         if (!myPrompt) {
             this.telemetryService.log('CodyVSCodeExtension:command:find:invalid')
@@ -213,16 +213,18 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
                 }
                 let label: string | undefined
                 let description: string | undefined
+                let slashCommand: string | undefined
 
                 if (command.slashCommand) {
                     label = command.slashCommand
                     description = command.name || name
+                    slashCommand = command.slashCommand
                 } else {
                     label = command.name || name
                     description = command.type === 'default' ? '' : command.type
                 }
 
-                return createQuickPickItem(label, description, label === '/edit')
+                return { label, description, slashCommand }
             })
 
             // Show the list of prompts to the user using a quick pick menu
@@ -237,13 +239,14 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
                 return
             }
 
-            const selectedCommandID = selectedPrompt.label
+            const selectedCommandID =
+                'slashCommand' in selectedPrompt ? selectedPrompt.slashCommand : selectedPrompt.label
             switch (true) {
                 case !selectedCommandID:
                     break
                 case selectedCommandID === menu_options.config.label:
                     return await vscode.commands.executeCommand('cody.settings.commands')
-                case selectedCommandID === menu_options.chat.label: {
+                case selectedCommandID === menu_options.chat.slashCommand: {
                     let input = userPrompt.trim()
                     if (input) {
                         return await vscode.commands.executeCommand('cody.action.chat', input)
@@ -252,8 +255,7 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
                     await vscode.commands.executeCommand('cody.chat.focus')
                     return await vscode.commands.executeCommand('cody.action.chat', input)
                 }
-
-                case selectedCommandID === menu_options.fix.label: {
+                case selectedCommandID === menu_options.fix.slashCommand: {
                     if (userPrompt.trim()) {
                         return await vscode.commands.executeCommand('cody.action.fixup', userPrompt)
                     }
@@ -269,7 +271,7 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
 
             await vscode.commands.executeCommand('cody.action.commands.exec', selectedCommandID)
         } catch (error) {
-            debug('CommandsController:commandQuickPicker', 'error', { verbose: error })
+            logError('CommandsController:commandQuickPicker', 'error', { verbose: error })
         }
     }
 
@@ -321,9 +323,9 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
                     break
             }
 
-            debug('CommandsController:promptsQuickPicker:selectedPrompt', promptTitle)
+            logDebug('CommandsController:promptsQuickPicker:selectedPrompt', promptTitle)
         } catch (error) {
-            debug('CommandsController:promptsQuickPicker', 'error', { verbose: error })
+            logError('CommandsController:promptsQuickPicker', 'error', { verbose: error })
         }
     }
 
@@ -342,7 +344,7 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
             return this.customCommandMenu()
         }
 
-        debug('CommandsController:customPrompts:menu', action)
+        logDebug('CommandsController:customPrompts:menu', action)
 
         switch (action) {
             case 'delete':
@@ -412,7 +414,7 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
         await this.custom.save(newCommand.title, newCommand.prompt)
         await this.refresh()
 
-        debug('CommandsController:updateUserCommandQuick:newPrompt:', 'saved', { verbose: newCommand })
+        logDebug('CommandsController:updateUserCommandQuick:newPrompt:', 'saved', { verbose: newCommand })
     }
 
     /**
@@ -499,7 +501,7 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
             )
         }
 
-        debug('CommandsController:fileWatcherInit', 'watchers created')
+        logDebug('CommandsController:fileWatcherInit', 'watchers created')
     }
 
     /**
@@ -517,6 +519,6 @@ export class CommandsController implements VsCodeCommandsController, vscode.Disp
         this.fileWatcherDisposables = []
         this.disposables = []
         this.myPromptsMap = new Map<string, CodyPrompt>()
-        debug('CommandsController:dispose', 'disposed')
+        logDebug('CommandsController:dispose', 'disposed')
     }
 }