+++ vscode/src/services/StatusBar.ts
@@ -6,16 +6,25 @@ import { getConfiguration } from '../configuration'
 
 import { FeedbackOptionItems } from './FeedbackOptions'
 
+interface StatusBarError {
+    title: string
+    description: string
+    onSelect?: () => void
+}
+
 export interface CodyStatusBar {
     dispose(): void
     startLoading(label: string): () => void
+    addError(error: StatusBarError): void
 }
 
 const DEFAULT_TEXT = '$(cody-logo-heavy)'
 const DEFAULT_TOOLTIP = 'Cody Settings'
 
 const QUICK_PICK_ITEM_CHECKED_PREFIX = '$(check) '
-const QUICK_PICK_ITEM_EMPTY_INDENT_PREFIX = '\u00A0\u00A0\u00A0\u00A0 '
+const QUICK_PICK_ITEM_EMPTY_INDENT_PREFIX = '\u00A0\u00A0\u00A0\u00A0\u00A0 '
+
+const ONE_HOUR = 60 * 60 * 1000
 
 export function createStatusBar(): CodyStatusBar {
     const statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right)
@@ -59,6 +68,23 @@ export function createStatusBar(): CodyStatusBar {
         const option = await vscode.window.showQuickPick(
             // These description should stay in sync with the settings in package.json
             [
+                ...(errors.length > 0
+                    ? [
+                          { label: 'notice', kind: vscode.QuickPickItemKind.Separator },
+                          ...errors.map(error => ({
+                              label: `$(alert) ${error.error.title}`,
+                              description: '',
+                              detail: QUICK_PICK_ITEM_EMPTY_INDENT_PREFIX + error.error.description,
+                              onSelect(): Promise<void> {
+                                  error.error.onSelect?.()
+                                  const index = errors.indexOf(error)
+                                  errors.splice(index)
+                                  rerender()
+                                  return Promise.resolve()
+                              },
+                          })),
+                      ]
+                    : []),
                 { label: 'enable/disable features', kind: vscode.QuickPickItemKind.Separator },
                 createFeatureToggle(
                     'Code Autocomplete',
@@ -129,11 +155,41 @@ export function createStatusBar(): CodyStatusBar {
     // TODO: Ensure the label is always set to the right value too.
     let openLoadingLeases = 0
 
+    const errors: { error: StatusBarError; createdAt: number }[] = []
+
+    function rerender(): void {
+        if (openLoadingLeases > 0) {
+            statusBarItem.text = '$(loading~spin)'
+        } else {
+            statusBarItem.text = DEFAULT_TEXT
+            statusBarItem.tooltip = DEFAULT_TOOLTIP
+        }
+
+        if (errors.length > 0) {
+            statusBarItem.backgroundColor = new vscode.ThemeColor('statusBarItem.warningBackground')
+            statusBarItem.tooltip = errors[0].error.title
+        } else {
+            statusBarItem.backgroundColor = new vscode.ThemeColor('statusBarItem.activeBackground')
+        }
+    }
+
+    // Clean up all errors after a certain time so they don't accumulate forever
+    function clearOutdatedErrors(): void {
+        const now = Date.now()
+        for (let i = errors.length - 1; i >= 0; i--) {
+            const error = errors[i]
+            if (now - error.createdAt >= ONE_HOUR) {
+                errors.splice(i, 1)
+            }
+        }
+        rerender()
+    }
+
     return {
         startLoading(label: string) {
             openLoadingLeases++
-            statusBarItem.text = '$(loading~spin)'
             statusBarItem.tooltip = label
+            rerender()
 
             let didClose = false
             return () => {
@@ -143,11 +199,13 @@ export function createStatusBar(): CodyStatusBar {
                 didClose = true
 
                 openLoadingLeases--
-                if (openLoadingLeases === 0) {
-                    statusBarItem.text = DEFAULT_TEXT
-                    statusBarItem.tooltip = DEFAULT_TOOLTIP
-                }
+                rerender()
             }
+        },
+        addError(error: StatusBarError) {
+            errors.push({ error, createdAt: Date.now() })
+            setTimeout(clearOutdatedErrors, ONE_HOUR)
+            rerender()
         },
         dispose() {
             statusBarItem.dispose()